fix: melhorias no código

+++ Backend/src/PopaDin.Bkd.Api/Dtos/Alert/AlertResponse.cs
@@ -5,7 +5,7 @@ public class AlertResponse
     public string Id { get; set; } = "";
     public int UserId { get; set; }
     public string Type { get; set; } = "";
-    public double Threshold { get; set; }
+    public decimal Threshold { get; set; }
     public string Channel { get; set; } = "";
     public bool Active { get; set; }
     public DateTime CreatedAt { get; set; }
