+++ Backend/src/PopaDin.Bkd.Api/Dtos/Alert/CreateAlertRequest.cs
@@ -5,6 +5,6 @@ namespace PopaDin.Bkd.Api.Dtos.Alert;
 public class CreateAlertRequest
 {
     public AlertType Type { get; set; }
-    public double Threshold { get; set; }
+    public decimal Threshold { get; set; }
     public string Channel { get; set; } = "";
 }
