+++ Backend/src/PopaDin.Bkd.Api/Dtos/Budget/BudgetResponse.cs
@@ -4,9 +4,8 @@ public class BudgetResponse
 {
     public int Id { get; set; }
     public string Name { get; set; } = "";
-    public double Goal { get; set; }
+    public decimal Goal { get; set; }
     public int UserId { get; set; }
-    // public UserResponseDTO User { get; set; }
     public DateTime? FinishAt { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set; }
