+++ Backend/src/PopaDin.Bkd.Api/Dtos/Budget/UpdateBudgetRequest.cs
@@ -3,6 +3,5 @@ namespace PopaDin.Bkd.Api.Dtos.Budget;
 public class UpdateBudgetRequest
 {
     public string Name { get; set; } = "";
-    public double Goal { get; set; }
-    // public DateTime FinishAt { get; set; }
+    public decimal Goal { get; set; }
 }
