+++ Backend/src/PopaDin.Bkd.Domain/Interfaces/Repositories/IRecordRepository.cs
@@ -1,5 +1,4 @@
 using PopaDin.Bkd.Domain.Models;
-using PopaDin.Bkd.Domain.Models.Record;
 
 namespace PopaDin.Bkd.Domain.Interfaces.Repositories;
 
@@ -7,7 +6,7 @@ public interface IRecordRepository
 {
     Task<Record> CreateRecordAsync(Record record, List<int> tagIds);
     Task<PaginatedResult<Record>> GetRecordsAsync(ListRecords listRecords);
-    Task<Record> FindRecordByIdAsync(decimal recordId, decimal userId);
+    Task<Record> FindRecordByIdAsync(int recordId, int userId);
     Task UpdateRecordAsync(Record record, List<int> tagIds);
-    Task DeleteRecordAsync(decimal recordId);
+    Task DeleteRecordAsync(int recordId);
 }
