+++ PopaDin.AlertService/Models/AlertRule.cs
@@ -11,7 +11,7 @@ public class AlertRule
 
     public int UserId { get; set; }
     public string Type { get; set; } = "";
-    public double Threshold { get; set; }
+    public decimal Threshold { get; set; }
     public string Channel { get; set; } = "";
     public bool Active { get; set; }
     public DateTime CreatedAt { get; set; }
